+++ src/compas_viewer/config.py
@@ -7,6 +7,7 @@ from typing import Literal
 
 from compas.colors import Color
 from compas_viewer.actions import change_viewmode
+from compas_viewer.actions import clear_scene
 from compas_viewer.actions import open_camera_settings_dialog
 from compas_viewer.actions import pan_view
 from compas_viewer.actions import rotate_view
@@ -102,7 +103,7 @@ class MenubarConfig(ConfigBase):
             {
                 "title": "Scene",
                 "items": [
-                    {"title": "Clear Scene", "action": lambda: print("Clear scene")},
+                    {"title": "Clear Scene", "action": clear_scene},
                     {"type": "separator"},
                     {"title": "Load Scene", "action": lambda: print("Load scene")},
                     {"title": "Save Scene", "action": lambda: print("Save scene")},