+++ src/compas_viewer/ui/mainwindow.py
@@ -1,10 +1,7 @@
-from PySide6 import QtWidgets
 from PySide6.QtWidgets import QMainWindow
-from PySide6.QtWidgets import QWidget
 
 
 class MainWindow:
-
     def __init__(self, title):
         self.title = title
         self.widget = QMainWindow()